docs: 주석, 명시적 형변환, toString 삭제 필요하지 않은 코드 부분 삭제

+++ src/main/java/com/carbonhater/co2zerobookmark/bookmark/model/dto/FolderHierarchyDto.java
@@ -45,7 +45,7 @@ public class FolderHierarchyDto {
         private String bookmarkUrl;
 
         public BookmarkDto(Bookmark bookmark) {
-            this.bookmarkId = (Long) bookmark.getBookmarkId();
+            this.bookmarkId = bookmark.getBookmarkId();
             this.bookmarkName = bookmark.getBookmarkName();
             this.bookmarkUrl = bookmark.getBookmarkUrl();
         }
